1. fox 添加系统版本号，版本号固定在bin文件0x1a0位置 2. fox neutron gl1000 core_hal.cpp 相同部分放入core_hal_stm32f4xx.cpp 3. 优化部分代码
