+++ hal/inc/variants/ant/pinmap_hal.h
@@ -67,9 +67,11 @@ STM32_Pin_Info* HAL_Pin_Map(void);
 * Pin mapping. Borrowed from Wiring
 */
 #define TOTAL_PINS             49
-#define TOTAL_ANALOG_PINS      6
+#define FIRST_DIGITAL_PIN      0
 #define TOTAL_DIGITAL_PINS     8
 #define FIRST_ANALOG_PIN       30
+#define TOTAL_ANALOG_PINS      6
+
 
 #define D0 0
 #define D1 1
