+++ hal/inc/variants/anytest/pinmap_hal.h
@@ -66,9 +66,10 @@ STM32_Pin_Info* HAL_Pin_Map(void);
 * Pin mapping. Borrowed from Wiring
 */
 #define TOTAL_PINS             47
-#define TOTAL_ANALOG_PINS      16
+#define FIRST_DIGITAL_PIN      0
 #define TOTAL_DIGITAL_PINS     26
 #define FIRST_ANALOG_PIN       30
+#define TOTAL_ANALOG_PINS      16
 
 #define D0  0
 #define D1  1
