+++ hal/inc/variants/fox/pinmap_hal.h
@@ -66,9 +66,10 @@ STM32_Pin_Info* HAL_Pin_Map(void);
 * Pin mapping. Borrowed from Wiring
 */
 #define TOTAL_PINS             37
-#define TOTAL_ANALOG_PINS      7
+#define FIRST_DIGITAL_PIN      0
 #define TOTAL_DIGITAL_PINS     8
 #define FIRST_ANALOG_PIN       30
+#define TOTAL_ANALOG_PINS      7
 
 #define D0 0
 #define D1 1
