+++ hal/inc/variants/l6/pinmap_hal.h
@@ -67,9 +67,10 @@ STM32_Pin_Info* HAL_Pin_Map(void);
 * Pin mapping. Borrowed from Wiring
 */
 #define TOTAL_PINS             31
-#define TOTAL_ANALOG_PINS      6
+#define FIRST_DIGITAL_PIN      0
 #define TOTAL_DIGITAL_PINS     24
 #define FIRST_ANALOG_PIN       9
+#define TOTAL_ANALOG_PINS      6
 
 #define D0 0
 #define D1 1
